requirements.txt
deepchem>=2.8,<3
tensorflow-cpu>=2.15,<2.17
scikit-learn>=1.4,<2
numpy>=1.26,<3
pandas>=2.2,<3
matplotlib>=3.8,<4
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
jinja2>=3.1,<4
joblib>=1.4,<2
python-multipart>=0.0.9,<1
rdkit-pypi>=2024.3.1
